backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
scikit-learn>=1.2.0
pandas>=1.5.0
joblib>=1.3.0
PyPDF2>=3.1.0
python-dateutil>=2.8.0
spacy>=3.5.0
email-validator>=2.0.0
pydantic>=2.0.0
